feat: arema main portal and video indexing (#64)
* feat: added some field hints #56

* feat: upload multiple files #56

* feat: markdown edition support #56

* feat: building fields added #56

* feat: fire classes #56

* fix: building material can have multiple types

* feat: clust results in frontend

* feat: entity page added to frontend (wip)

* feat: entity page added to frontend (wip)

* feat: search page state in store (wip)

* feat: fire properties have low/high and input suggestions

* feat: markdown guide link

* feat: use search service to retrieve document from éain portal

* feat: FTKG font applied

* feat: physical characteristics

* feat: search panel, hash path

* feat: index page style

* feat: relations are indexed

* feat: added building relation with tc via be

* feat: document images carousel

* feat: display images

* feat: file can be a url, youtube/vimeo videos handled in a player

* fix: ensure taxos before getting tags labels

* feat: search link added in doc page

* feat: select all terms if none are selected, otherwise clear selections

* feat: term btns are smaller than voc ones

* fix: taxo labels typo

* fix: fire norms

* feat: prepare frontend for it lang

* feat: map's list is scrollable

* feat: styling

* chore: tr

* chore: close drawer on route update

* feat: videos indexed and displayed in a search interface #63

* fix: videos page title

* feat: extract viedo links from markdown text #63

* fix: text input model type

+++ backend/api/services/search.py
@@ -1,65 +1,21 @@
-from datetime import datetime
+from abc import ABC, abstractmethod
+import copy
+import re
+from markdown_it import MarkdownIt
 from elasticsearch import Elasticsearch
 from api.config import config
 from api.models.query import SearchResult
 
-ANALYZED_FIELDS = ["name", "description", "article_top",
-                   "article_bottom", "side_note", "address"]
+ENTITY_ANALYZED_FIELDS = ["name", "description", "article_top",
+                          "article_bottom", "side_note", "address"]
+
+VIDEO_ANALYZED_FIELDS = ["name", "legend"]
 
 
-class IndexService:
-    def __init__(self, index_name: str = "arema"):
+class IndexService(ABC):
+    def __init__(self, index_name: str):
         self.client = Elasticsearch(config.ES_URL)
         self.index_name = index_name
-
-    def addEntity(self, entity_type: str, entity, tags: list[str]):
-        """Add a new entity to the index
-
-        Args:
-            entity_type (str): The class of the entity
-            entity (_type_): The entity object to add
-            tags (list[str]): The associated tags
-        """
-        doc_id = f"{entity_type}:{entity.id}"
-        doc = entity.model_dump()
-        doc["entity_type"] = entity_type
-        doc["tags"] = tags
-        if "long" in doc and "lat" in doc:
-            doc["location"] = {"lat": doc["lat"], "lon": doc["long"]}
-        self._addDocument(doc_id, doc)
-
-    def updateEntity(self, entity_type: str, entity, tags: list[str]):
-        """Update an entity of the index
-
-        Args:
-            entity_type (str): The class of the entity
-            entity (_type_): The entity object to add
-            tags (list[str]): The associated tags
-        """
-        doc_id = f"{entity_type}:{entity.id}"
-        doc = entity.model_dump()
-        doc["entity_type"] = entity_type
-        doc["tags"] = tags
-        self._updateDocument(doc_id, doc)
-
-    def deleteEntity(self, entity_type: str, entity_id: int):
-        """Delete an entity from the index
-
-        Args:
-            entity_type (str): The class of the entity
-            entity_id (int): The id of the entity
-        """
-        doc_id = f"{entity_type}:{entity_id}"
-        self._deleteDocument(doc_id)
-
-    def deleteEntities(self, entity_type: str):
-        """Delete all entities of a given type
-
-        Args:
-            entity_type (str): The class of the entities
-        """
-        query = {"query": {"term": {"entity_type": entity_type}}}
-        self._deleteDocuments(query)
 
     def refreshIndex(self):
         """Refresh the index to make changes visible
@@ -103,12 +59,75 @@ class IndexService:
 
     def _deleteDocument(self, doc_id):
         self._ensureIndex()
-        self.client.delete(index=self.index_name, id=doc_id)
+        if self.client.exists(index=self.index_name, id=doc_id):
+            self.client.delete(index=self.index_name, id=doc_id)
 
     def _deleteDocuments(self, query: dict):
         self._ensureIndex()
         self.client.delete_by_query(
             index=self.index_name, body=query, ignore_unavailable=True)
+
+    @abstractmethod
+    def _ensureIndex(self):
+        """Ensure the index exists and has the correct mapping
+        """
+        pass
+
+
+class EntityIndexService(IndexService):
+    def __init__(self):
+        super().__init__("entities")
+
+    def addEntity(self, entity_type: str, entity, tags: list[str], relates_to: list[str] = []):
+        """Add a new entity to the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+            tags (list[str]): The associated tags
+        """
+        doc_id = f"{entity_type}:{entity.id}"
+        doc = entity.model_dump()
+        doc["entity_type"] = entity_type
+        doc["tags"] = tags
+        doc["relates_to"] = relates_to
+        if "long" in doc and "lat" in doc:
+            doc["location"] = {"lat": doc["lat"], "lon": doc["long"]}
+        self._addDocument(doc_id, doc)
+
+    def updateEntity(self, entity_type: str, entity, tags: list[str], relates_to: list[str] = []):
+        """Update an entity of the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+            tags (list[str]): The associated tags
+        """
+        doc_id = f"{entity_type}:{entity.id}"
+        doc = entity.model_dump()
+        doc["entity_type"] = entity_type
+        doc["tags"] = tags
+        doc["relates_to"] = relates_to
+        self._updateDocument(doc_id, doc)
+
+    def deleteEntity(self, entity_type: str, entity_id: int):
+        """Delete an entity from the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity_id (int): The id of the entity
+        """
+        doc_id = f"{entity_type}:{entity_id}"
+        self._deleteDocument(doc_id)
+
+    def deleteEntities(self, entity_type: str):
+        """Delete all entities of a given type
+
+        Args:
+            entity_type (str): The class of the entities
+        """
+        query = {"query": {"term": {"entity_type": entity_type}}}
+        self._deleteDocuments(query)
 
     def _ensureIndex(self):
         if not self.client.indices.exists(index=self.index_name):
@@ -136,6 +155,9 @@ class IndexService:
                         "tags": {
                             "type": "keyword"
                         },
+                        "relates_to": {
+                            "type": "keyword"
+                        },
                         "entity_type": {
                             "type": "keyword"
                         },
@@ -145,7 +167,7 @@ class IndexService:
                     }
                 }
             }
-            for field in ANALYZED_FIELDS:
+            for field in ENTITY_ANALYZED_FIELDS:
                 configuration["mappings"]["properties"][field] = {
                     "type": "text",
                     "analyzer": "ngram_analyzer"
@@ -154,3 +176,270 @@ class IndexService:
                 index=self.index_name, body=configuration)
             self.client.indices.refresh(index=self.index_name)
             self.client.cluster.health(wait_for_status="yellow")
+
+
+class VideoIndexService(IndexService):
+    def __init__(self):
+        super().__init__("videos")
+
+    def isVideoUrl(self, url: str):
+        """Check if a given url is a video
+
+        Args:
+            url (str): The url to check
+
+        Returns:
+            bool: True if the url is a video
+        """
+        return url is not None and ("youtube.com" in url or "youtu.be" in url or "vimeo.com" in url or "srf.ch/play/tv" in url or "rts.ch/play/tv" in url)
+
+    def addVideos(self, entity_type: str, entity, tags: list[str]):
+        """Add videos of a new entity to the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+        """
+        for doc in self._getVideos(entity_type, entity, tags):
+            self._addDocument(doc["id"], doc)
+
+    def updateVideos(self, entity_type: str, entity, tags: list[str]):
+        """Update videos of an entity of the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+            tags (list[str]): The associated tags
+        """
+        for doc in self._getVideos(entity_type, entity, tags):
+            self._updateDocument(doc["id"], doc)
+
+    def deleteVideos(self, entity_type: str, entity_id: int):
+        """Delete videos of an entity from the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity_id (int): The id of the entity
+        """
+        parent_id = f"{entity_type}:{entity_id}"
+        query = {"query": {"term": {"parent_id": parent_id}}}
+        self._deleteDocument(query)
+
+    def deleteVideos(self, entity_type: str):
+        """Delete all videos of a given type
+
+        Args:
+            entity_type (str): The class of the entities
+        """
+        query = {"query": {"term": {"entity_type": entity_type}}}
+        self._deleteDocuments(query)
+
+    def _getVideos(self, entity_type: str, entity, tags: list[str]):
+        """Get videos of an entity
+
+        Args:
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+            tags (list[str]): The associated tags
+
+        Returns:
+            list[dict]: The list of video documents 
+        """
+        docs = []
+        docs += self._getVideosFromFiles(entity_type, entity, tags)
+        for field in ["external_links", "article_top", "article_bottom", "side_note"]:
+            docs += self._getVideosFromField(field, entity_type, entity, tags)
+        return docs
+
+    def _getVideosFromFiles(self, entity_type: str, entity, tags: list[str]):
+        """Get videos of an entity from the files attribute
+
+        Args:
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+            tags (list[str]): The associated tags
+
+        Returns:
+            list[dict]: The list of video documents 
+        """
+        if not entity.files:
+            return []
+        docs = []
+        parent_id = f"{entity_type}:{entity.id}"
+        for idx, file in enumerate(entity.files):
+            doc_id = f"{entity_type}:{entity.id}:file:{idx}"
+            if "url" in file and self.isVideoUrl(file["url"]):
+                doc = copy.deepcopy(file)
+                doc["id"] = doc_id
+                doc["name"] = entity.name
+                doc["entity_type"] = entity_type
+                doc["parent_id"] = parent_id
+                doc["tags"] = tags
+                docs.append(doc)
+        return docs
+
+    def _getVideosFromField(self, field, entity_type: str, entity, tags: list[str]):
+        """Get videos of an entity from the files attribute
+
+        Args:
+            field (str): The field to search for videos
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+            tags (list[str]): The associated tags
+
+        Returns:
+            list[dict]: The list of video documents 
+        """
+        text = getattr(entity, field)
+        if not text:
+            return []
+        docs = []
+        parent_id = f"{entity_type}:{entity.id}"
+        links = self.extract_all_links_with_text(text)
+        for idx, link in enumerate(links):
+            doc_id = f"{entity_type}:{entity.id}:{field}:{idx}"
+            if "url" in link and self.isVideoUrl(link["url"]):
+                doc = link
+                doc["id"] = doc_id
+                doc["name"] = entity.name
+                doc["entity_type"] = entity_type
+                doc["parent_id"] = parent_id
+                doc["tags"] = tags
+                docs.append(doc)
+        return docs
+
+    def extract_all_links_with_text(self, markdown_text):
+        md = MarkdownIt()
+        tokens = md.parse(markdown_text)
+        links = []
+
+        # Extract Markdown links
+        for i, token in enumerate(tokens):
+            if token.type == "link_open":
+                href = [attr[1]
+                        for attr in token.attrs if attr[0] == "href"][0]
+                if i + 1 < len(tokens) and tokens[i + 1].type == "text":
+                    legend = tokens[i + 1].content
+                    links.append({"legend": legend, "url": href})
+
+        # Extract plain URLs
+        plain_url_pattern = r'(https?://[^\s]+)'
+        plain_urls = re.findall(plain_url_pattern, markdown_text)
+
+        for url in plain_urls:
+            # Ensure plain URLs are not duplicates of Markdown links
+            if not any(link["url"] == url for link in links):
+                links.append({"legend": None, "url": url})
+
+        return links
+
+    def _ensureIndex(self):
+        if not self.client.indices.exists(index=self.index_name):
+            configuration = {
+                "settings": {
+                    "analysis": {
+                        "analyzer": {
+                            "ngram_analyzer": {
+                                "type": "custom",
+                                "tokenizer": "standard",
+                                "filter": ["lowercase", "ngram_filter"]
+                            }
+                        },
+                        "filter": {
+                            "ngram_filter": {
+                                "type": "ngram",
+                                "min_gram": 3,  # minimum n-gram length
+                                "max_gram": 4   # maximum n-gram length
+                            }
+                        }
+                    }
+                },
+                "mappings": {
+                    "properties": {
+                        "tags": {
+                            "type": "keyword"
+                        },
+                        "parent_id": {
+                            "type": "keyword"
+                        },
+                        "entity_type": {
+                            "type": "keyword"
+                        }
+                    }
+                }
+            }
+            for field in VIDEO_ANALYZED_FIELDS:
+                configuration["mappings"]["properties"][field] = {
+                    "type": "text",
+                    "analyzer": "ngram_analyzer"
+                }
+            self.client.indices.create(
+                index=self.index_name, body=configuration)
+            self.client.indices.refresh(index=self.index_name)
+            self.client.cluster.health(wait_for_status="yellow")
+
+
+class EntityIndexer:
+    def __init__(self):
+        self.entityIndexService = EntityIndexService()
+        self.videoIndexService = VideoIndexService()
+
+    def addEntity(self, entity_type: str, entity, tags: list[str], relates_to: list[str] = []):
+        """Add a new entity to the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+            tags (list[str]): The associated tags
+        """
+        self.entityIndexService.addEntity(
+            entity_type, entity, tags, relates_to)
+        self.videoIndexService.addVideos(entity_type, entity, tags)
+
+    def updateEntity(self, entity_type: str, entity, tags: list[str], relates_to: list[str] = []):
+        """Update an entity of the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity (_type_): The entity object to add
+            tags (list[str]): The associated tags
+        """
+        self.entityIndexService.updateEntity(
+            entity_type, entity, tags, relates_to)
+        self.videoIndexService.updateVideos(entity_type, entity, tags)
+
+    def deleteEntity(self, entity_type: str, entity_id: int):
+        """Delete an entity from the index
+
+        Args:
+            entity_type (str): The class of the entity
+            entity_id (int): The id of the entity
+        """
+        self.entityIndexService.deleteEntity(entity_type, entity_id)
+        self.videoIndexService.deleteVideos(entity_type, entity_id)
+
+    def deleteEntities(self, entity_type: str):
+        """Delete all entities of a given type
+
+        Args:
+            entity_type (str): The class of the entities
+        """
+        self.entityIndexService.deleteEntities(entity_type)
+        self.videoIndexService.deleteVideos(entity_type)
+
+    def deleteIndex(self):
+        """Drop the index, its mapping definitions and all its data
+        """
+        self.entityIndexService.deleteIndex()
+        self.videoIndexService.deleteIndex()
+
+
+class SearchService:
+
+    @classmethod
+    def fromIndex(cls, index: str):
+        if index == "entities":
+            return EntityIndexService()
+        if index == "videos":
+            return VideoIndexService()
+        raise ValueError(f"Unknown index: {index}")
